+++ alpharaw/viz/xic_plot.py
@@ -7,13 +7,16 @@ import plotly.graph_objects as go
 import plotly.express as px
 from plotly.subplots import make_subplots
 
-from alphatims.bruker import (
-    TimsTOF, 
+from .plot_utils import (
+    plot_line
 )
 
-from .plot_utils import (
-    plot_line_fast
+from alphabase.constants.atom import MASS_ISOTOPE
+from alpharaw.match.spec_finder import (
+    find_spec_idxes,
+    find_multinotch_spec_idxes,
 )
+from alpharaw.match.match_utils import match_batch_spec
 
 class XIC_Plot():
     # hovermode = "x" | "y" | "closest" | False | "x unified" | "y unified"
@@ -24,29 +27,34 @@ class XIC_Plot():
     theme_template='plotly_white'
     ms1_ppm = 20.0
     ms2_ppm = 20.0
-    rt_sec_win = 30.0
+    rt_sec_win = 60.0
     plot_rt_unit:str = "minute"
-    im_win = 0.05
     fig:go.Figure = None
+    isotope_min_abundance=0.01,
 
-    #list of XIC_Trace objects
+    # list of XIC_Trace objects
     traces:list = []
 
+    """
+    Only apply for 3D-MS, i.e. without mobility
+    """
+
     def plot(self, 
-        tims_data:TimsTOF,
+        spectrum_df:pd.DataFrame,
+        peak_df:pd.DataFrame,
         query_df:pd.DataFrame,
-        view_dim:typing.Literal["rt","im"]="rt",
         title:str="",
+        peak_area_for_each_query=False,
     ):
-        rt_sec = query_df['rt_sec'].values[0]
-        if "im" not in query_df.columns:
-            im = 0.0
+        if "rt_sec" in query_df.columns:
+            rt_sec = query_df.rt_sec.values[0]
         else:
-            im = query_df["im"].values[0]
-        if "precursor_mz" not in query_df.columns:
-            precursor_mz = 0.0
-        else:
-            precursor_mz = query_df.precursor_mz.values[0]
+            rt_sec = query_df.rt.values[0]*60
+        
+        (
+            precursor_left_mz, precursor_right_mz
+        ) = self._get_precursor_mz_range(query_df)
+        
         query_masses = query_df.mz.values
         if "intensity" in query_df.columns:
             query_intensities = query_df.intensity.values
@@ -59,64 +67,111 @@ class XIC_Plot():
         else:
             marker_colors = query_df.color.values
 
+        if peak_area_for_each_query:
+            self.get_peak_area(
+                spectrum_df, peak_df, query_df
+            )
+
         return self.plot_query_masses(
-            tims_data=tims_data,
+            spectrum_df=spectrum_df,
+            peak_df=peak_df,
             query_masses=query_masses,
             query_ion_names=ion_names,
             query_rt_sec=rt_sec,
-            query_im=im,
-            precursor_mz=precursor_mz,
+            precursor_left_mz=precursor_left_mz,
+            precursor_right_mz=precursor_right_mz,
             marker_colors=marker_colors,
-            view_dim=view_dim,
             query_intensities=query_intensities,
             title=title,
         )
+    
+    def _get_precursor_mz_range(
+        self, query_df,
+    ):
+        if "precursor_mz" not in query_df.columns:
+            precursor_left_mz = -1.0
+            precursor_right_mz = -1.0
+        else:
+            precursor_left_mz = query_df.precursor_mz.values[0]*(1-self.ms1_ppm*1e-6)
+            precursor_right_mz = query_df.precursor_mz.values[0]*(1+self.ms1_ppm*1e-6)
+        
+            for iso in range(10):
+                if f"precursor_i_{iso}" not in query_df.columns:
+                    break
+            if iso > 0:
+                mono_idx = query_df.precursor_mono_idx.values[0]
+                charge = query_df.precursor_charge.values[0]
+                precursor_mz = query_df.precursor_mz.values[0]
+                precursor_left_mz = precursor_mz-mono_idx*MASS_ISOTOPE/charge
+                for i in range(iso-1,-1,-1):
+                    if query_df[f"precursor_i_{i}"].values[0] >= self.isotope_min_abundance:
+                        precursor_right_mz = precursor_mz+(i-mono_idx)*MASS_ISOTOPE/charge
+                        break
+        return precursor_left_mz, precursor_right_mz
+    
+    def get_peak_area(self, 
+        spectrum_df:pd.DataFrame,
+        peak_df:pd.DataFrame,
+        query_df:pd.DataFrame,
+        query_start_rt_sec:float=None,
+        query_stop_rt_sec:float=None,
+        precursor_left_mz:float=None,
+        precursor_right_mz:float=None,
+    ):
+        if query_start_rt_sec is None:
+            query_start_rt_sec = query_df.rt_sec.values[0] - self.rt_sec_win/2
+            query_stop_rt_sec = query_df.rt_sec.values[0] + self.rt_sec_win/2
+        if precursor_left_mz is None:
+            (
+                precursor_left_mz, precursor_right_mz
+            ) = self._get_precursor_mz_range(query_df)
 
-    def _init_plot(self, rows=1, view_dim='rt'):
-        self.fig = make_subplots(
-            rows=rows, 
-            cols=1,
-            shared_xaxes=True,
-            x_title=f'RT ({self.plot_rt_unit})' if view_dim == 'rt' else 'Mobility',
-            y_title='intensity',
-            vertical_spacing=0.2/rows,
+        mass_tols = query_df.mz.values*1e-6*(
+            self.ms1_ppm if precursor_left_mz<=0 else self.ms2_ppm
         )
-        self.traces = [
-            XIC_Trace(fig=self.fig, row=i+1) 
-            for i in range(rows)
-        ]
+
+        query_df["peak_area"] = get_peak_area(
+            spectrum_df=spectrum_df,
+            peak_df=peak_df,
+            query_masses=query_df.mz.values,
+            mass_tols=mass_tols,
+            query_start_rt_sec=query_start_rt_sec,
+            query_stop_rt_sec=query_stop_rt_sec,
+            precursor_left_mz=precursor_left_mz,
+            precursor_right_mz=precursor_right_mz,
+        )
+        return query_df
 
     def plot_query_masses(self,
-        tims_data:TimsTOF,
+        spectrum_df:pd.DataFrame,
+        peak_df:pd.DataFrame,
         query_masses:np.ndarray,
         query_ion_names:typing.List[str],
         query_rt_sec:float, 
-        query_im:float,
-        precursor_mz:float,
+        precursor_left_mz:float,
+        precursor_right_mz:float,
         marker_colors:list = None,
-        view_dim:typing.Literal["rt","im"]="rt",
         query_intensities:np.ndarray = None,
         title="",
+        ith_plot=0,
     ):
-        self._init_plot(rows=1, view_dim=view_dim)
+        self._init_plot(rows=1)
         mass_tols = query_masses*1e-6*(
-            self.ms1_ppm if precursor_mz == 0 else self.ms2_ppm
+            self.ms1_ppm if precursor_left_mz<0 else self.ms2_ppm
         )
         if marker_colors is None:
             marker_colors = self._get_color_set(len(query_masses))
-        self.traces[0].add_traces(
-            tims_data=tims_data,
+        self.traces[ith_plot].add_traces(
+            spectrum_df=spectrum_df,
+            peak_df=peak_df,
             query_masses=query_masses,
             mass_tols=mass_tols,
             ion_names=query_ion_names,
             marker_colors=marker_colors,
-            query_rt_sec=query_rt_sec,
-            query_im=query_im,
-            precursor_mz=precursor_mz,
-            precursor_mz_tol=precursor_mz*1e-6*self.ms1_ppm,
-            view_dim=view_dim,
-            rt_sec_win=self.rt_sec_win,
-            im_win=self.im_win,
+            query_start_rt_sec=query_rt_sec-self.rt_sec_win/2,
+            query_stop_rt_sec=query_rt_sec+self.rt_sec_win/2,
+            precursor_left_mz=precursor_left_mz,
+            precursor_right_mz=precursor_right_mz,
             query_intensities=query_intensities,
         )
         self.fig.update_layout(
@@ -131,6 +186,23 @@ class XIC_Plot():
             showlegend=True,
         )
         return self.fig
+    
+    def _init_plot(self, rows=1):
+        self.fig = make_subplots(
+            rows=rows, 
+            cols=1,
+            shared_xaxes=True,
+            x_title=f'RT ({self.plot_rt_unit})',
+            y_title='intensity',
+            vertical_spacing=0.2/rows,
+        )
+        self.traces:typing.List[XIC_Trace] = [
+            XIC_Trace(
+                fig=self.fig, row=i+1, 
+                plot_rt_unit=self.plot_rt_unit,
+            ) 
+            for i in range(rows)
+        ]
 
     def _get_color_set(self, n_query):
         if n_query <= len(
@@ -165,45 +237,38 @@ class XIC_Trace():
         self.plot_rt_unit = plot_rt_unit
 
     def add_traces(self, 
-        tims_data:TimsTOF,
+        spectrum_df:pd.DataFrame,
+        peak_df:pd.DataFrame,
         query_masses:np.ndarray,
         mass_tols:np.ndarray,
         ion_names:typing.List[str],
         marker_colors:typing.List,
-        query_rt_sec:float, 
-        query_im:float,
-        precursor_mz:float = 0.0,
-        precursor_mz_tol:float = 0.02,
-        view_dim:typing.Literal["rt","im"]="rt",
-        rt_sec_win = 30.0,
-        im_win = 0.05,
+        query_start_rt_sec:float, 
+        query_stop_rt_sec:float, 
+        precursor_left_mz:float = -1.0,
+        precursor_right_mz:float = -1.0,
         query_intensities:np.ndarray = None,
     )->go.Figure:
-        """Add traces for the query_masses.
 
-        Args:
-            tims_data (TimsTOF): AlphaTims TimsTOF object.
-            query_masses (np.ndarray): Query masses.
-            ion_names (typing.List[str]): Ion names for query_masses.
-            marker_colors (typing.List): Colors for each query mass.
-            query_rt_sec (float): Query RT in seconds.
-            query_im (float): Query ion mobility.
-            precursor_mz (float, optional): Precursor mz, 0 means it is MS1. Defaults to 0.0.
-            query_intensities (np.ndarray, optional): Query intensities. Defaults to None.
+        spec_idxes = get_spec_idxes_from_df(
+            spectrum_df,
+            query_start_rt_sec=query_start_rt_sec,
+            query_stop_rt_sec=query_stop_rt_sec,
+            precursor_left_mz=precursor_left_mz,
+            precursor_right_mz=precursor_right_mz,
+        )
 
-        Returns:
-            go.Figure: self.fig.
-        """
-        (
-            rt_slice, im_slice, prec_mz_slice, view_indices
-        ) = get_plotting_slices(
-            tims_data=tims_data, 
-            rt_sec=query_rt_sec, 
-            rt_sec_win=rt_sec_win,
-            im=query_im, im_win=im_win,
-            precursor_mz=precursor_mz, 
-            precursor_mz_tol=precursor_mz_tol,
-            view_dim=view_dim,
+        rt_values = spectrum_df.rt.values[spec_idxes]
+        if self.plot_rt_unit == "second":
+            rt_values = rt_values * 60
+        matched_mzs, matched_intens = match_batch_spec(
+            spec_idxes, 
+            peak_mzs=peak_df.mz.values,
+            peak_intens=peak_df.intensity.values,
+            peak_start_idxes=spectrum_df.peak_start_idx.values,
+            peak_stop_idxes=spectrum_df.peak_stop_idx.values,
+            query_mzs=query_masses,
+            query_mz_tols=mass_tols,
         )
 
         if query_intensities is None:
@@ -211,140 +276,103 @@ class XIC_Trace():
         else:
             query_intensities /= query_intensities.max()
 
-        for (
-            ion_name, query_mass, query_inten,
-            marker_color, mass_tol
-        ) in zip(
-            ion_names, query_masses, query_intensities,
-            marker_colors, mass_tols
-        ):
+        for i_query, (
+            ion_name, query_mass, 
+            query_inten, marker_color
+        ) in enumerate(zip(
+            ion_names, query_masses, 
+            query_intensities, marker_colors
+        )):
+            if (matched_intens[:,i_query]==0).all(): continue
             self._add_one_trace(
-                tims_data=tims_data,
-                query_mass=query_mass, 
-                mass_tol=mass_tol, 
-                rt_slice=rt_slice, im_slice=im_slice,
-                prec_mz_slice=prec_mz_slice, 
-                view_indices=view_indices,
-                view_dim=view_dim,
+                rt_values, matched_intens[:,i_query],
                 label=self.label_format.format(ion_name=ion_name, mz=query_mass),
                 legend_group=self.legend_group.format(ion_name=ion_name),
                 marker_color=marker_color,
             )
             if query_inten > 0:
                 self._add_one_trace(
-                    tims_data=tims_data,
-                    query_mass=query_mass, 
-                    mass_tol=mass_tol, 
-                    rt_slice=rt_slice, im_slice=im_slice,
-                    prec_mz_slice=prec_mz_slice, 
-                    view_indices=view_indices,
-                    view_dim=view_dim,
+                    rt_values, -query_inten*matched_intens[:,i_query],
                     label=self.label_format.format(ion_name=ion_name, mz=query_mass),
                     legend_group=self.legend_group.format(ion_name=ion_name),
                     marker_color=marker_color,
-                    intensity_scale=-query_inten,
                 )
-
     def _add_one_trace(self,
-        tims_data:TimsTOF,
-        query_mass:float,
-        mass_tol:float,
-        rt_slice:slice, im_slice:slice, 
-        prec_mz_slice:slice,
-        view_indices:np.ndarray,
-        view_dim:str,
+        rt_values:np.ndarray,
+        matched_intens:np.ndarray,
         label:str, 
         legend_group:str, 
         marker_color:str,
-        intensity_scale:float=1.0,
     ):
-        frag_indices = tims_data[
-            rt_slice,
-            im_slice,
-            prec_mz_slice,
-            slice(
-                query_mass - mass_tol, 
-                query_mass + mass_tol,
-            ),
-            'raw'
-        ]
-        if len(frag_indices) == 0: return
         self.fig.add_trace(
-            plot_line_fast(
-                tims_data, 
-                frag_indices,
-                view_indices,
+            plot_line(
+                rt_values, 
+                matched_intens,
                 name=label,
                 legend_group=legend_group,
                 marker_color=marker_color, 
-                view_dim=view_dim,
-                intensity_scale=intensity_scale
+                x_text="RT",
             ),
             row=self.row, col=self.col,
         )
 
-def get_plotting_slices(
-    tims_data: TimsTOF,
-    rt_sec:float, rt_sec_win:float=30.0,
-    im:float = 0., im_win:float = 0.05,
-    precursor_mz:float = 0.0,
-    precursor_mz_tol:float = 0.02,
-    view_dim:str = "rt",
+def get_spec_idxes_from_df(
+    spectrum_df:pd.DataFrame,
+    query_start_rt_sec,
+    query_stop_rt_sec,
+    precursor_left_mz,
+    precursor_right_mz,
 ):
-    """
-    Get plotting slices for target queries in TimsTOF data.
-    Args:
-        tims_data (TimsTOF): AlphaTims TimsTOF object.
-        rt_sec (float): Query RT in seconds.
-        rt_sec_win (float, optional): Query RT window in seconds. Defaults to 30.0.
-        im (float, optional): Query ion mobility, 0 means no mobility dimension. Defaults to 0..
-        im_win (float, optional): Ion mobility window. Defaults to 0.05.
-        precursor_mz (float, optional): Precursor mz, 0 means it is MS1. Defaults to 0.0.
-        ppm (float, optional): PPM tolerance for `precursor_mz`. Defaults to 20.0.
-        view_dim (str, optional): View dimension, "rt" or "im". Defaults to "rt"
-    """
-    rt_slice = slice(
-        rt_sec - rt_sec_win/2, 
-        rt_sec + rt_sec_win/2
-    )
-    im_slice = slice(
-        im - im_win/2, 
-        im + im_win/2
-    )
-
-    if precursor_mz == 0:
-        prec_mz_slice = 0
-        raw_indices = tims_data[
-            rt_slice,
-            im_slice,
-            0,
-            :,
-            'raw'
-        ]
-    else:
-        prec_mz_slice = slice(
-            precursor_mz - precursor_mz_tol,
-            precursor_mz + precursor_mz_tol,
+    if "multinotch" in spectrum_df.columns:
+        return find_multinotch_spec_idxes(
+            spectrum_df.rt.values,
+            spectrum_df.multinotch.values,
+            query_start_rt=query_start_rt_sec/60,
+            query_stop_rt=query_stop_rt_sec/60,
+            query_left_mz=precursor_left_mz,
+            query_right_mz=precursor_right_mz
         )
-        raw_indices = tims_data[
-            rt_slice,
-            im_slice,
-            prec_mz_slice,
-            :,
-            'raw'
-        ]
-
-    if view_dim == 'rt':
-        view_indices = np.sort(np.array(list(set(
-            tims_data.convert_from_indices(
-            raw_indices,
-            return_frame_indices=True
-        )['frame_indices'])), dtype=np.int64))
     else:
-        view_indices = np.sort(np.array(list(set(
-            tims_data.convert_from_indices(
-            raw_indices,
-            return_scan_indices=True
-        )['scan_indices'])), dtype=np.int64))
+        return find_spec_idxes(
+            spectrum_df.rt.values,
+            spectrum_df.isolation_lower_mz.values,
+            spectrum_df.isolation_upper_mz.values,
+            query_start_rt=query_start_rt_sec/60,
+            query_stop_rt=query_stop_rt_sec/60,
+            query_left_mz=precursor_left_mz,
+            query_right_mz=precursor_right_mz
+        )
+    
+def get_peak_area(
+    spectrum_df:pd.DataFrame,
+    peak_df:pd.DataFrame,
+    query_masses:np.ndarray,
+    mass_tols:np.ndarray,
+    query_start_rt_sec:float, 
+    query_stop_rt_sec:float,
+    precursor_left_mz:float = -1.0,
+    precursor_right_mz:float = -1.0,
+)->np.ndarray:
+    spec_idxes = get_spec_idxes_from_df(
+        spectrum_df,
+        query_start_rt_sec=query_start_rt_sec,
+        query_stop_rt_sec=query_stop_rt_sec,
+        precursor_left_mz=precursor_left_mz,
+        precursor_right_mz=precursor_right_mz,
+    )
 
-    return rt_slice, im_slice, prec_mz_slice, view_indices
+    rt_values = spectrum_df.rt.values[spec_idxes]
+
+    matched_mzs, matched_intens = match_batch_spec(
+        spec_idxes, 
+        peak_mzs=peak_df.mz.values,
+        peak_intens=peak_df.intensity.values,
+        peak_start_idxes=spectrum_df.peak_start_idx.values,
+        peak_stop_idxes=spectrum_df.peak_stop_idx.values,
+        query_mzs=query_masses,
+        query_mz_tols=mass_tols,
+    )
+
+    return np.trapz(
+        y=matched_intens, x=rt_values*60, axis=0
+    )
